Move error handlers to HttpClientException
